+++ src/components/KanbanComponent/KanbanComponent.tsx
@@ -19,56 +19,7 @@ import { KanbanTeamsComponent } from './KanbanTeamsComponent';
 import { githubAuthApiRef, useApi } from '@backstage/core-plugin-api';
 import { KanbanColumnBody } from './KanbanBodyComponent';
 
-export type CommentType = {
-  author: {
-    login: string;
-  };
-  body: string;
-  createdAt: string;
-  updatedAt: string;
-};
-
-export type ReviewType = {
-  author: {
-    login: string;
-  };
-  body: string;
-  comments: {
-    nodes: CommentType[];
-  };
-  createdAt: string;
-  state: string;
-  updatedAt: string;
-};
-
-export type LabelType = {
-  color: string;
-  name: string;
-};
-
-export type PRType = {
-  author: {
-    login: string;
-  };
-  body: string;
-  createdAt: string;
-  number: number;
-  labels: {
-    nodes: LabelType[];
-  };
-  reviews: {
-    nodes: ReviewType[];
-  };
-  state: string;
-  title: string;
-  updatedAt: string;
-  url: string;
-};
-
-export type RepoType = {
-  repository: string;
-  data: PRType[];
-};
+import { RepoType, PRType } from './KanbanTypes';
 
 export const KanbanComponent = () => {
   const [query, setQuery] = useState<PRType>();
